requirements.txt
psycopg
spacy
python -m spacy download en_core_web_sm